Trait validation::ValidatedField

pub trait ValidatedField<T>where
    Self: Sized,{
    fn validate(
        input: Option<T>,
        field_name: &str,
        parent_name: Option<&str>
    ) -> Result<Self, ValidationError>;

    fn validate_optional(
        input: Option<T>,
        field_name: &str,
        parent_name: Option<&str>
    ) -> Result<Option<Self>, ValidationError> { ... }
    fn validate_for_creation(
        input: Option<T>,
        field_name: &str,
        parent_name: Option<&str>
    ) -> Result<Self, ValidationError> { ... }
}

Required Methods

fn validate(
input: Option<T>,
field_name: &str,
parent_name: Option<&str>
) -> Result<Self, ValidationError>

Provided Methods

fn validate_optional(
input: Option<T>,
field_name: &str,
parent_name: Option<&str>
) -> Result<Option<Self>, ValidationError>

Default implementation returns immediately when input is None. Otherwise, calls Self::validate.

fn validate_for_creation(
input: Option<T>,
field_name: &str,
parent_name: Option<&str>
) -> Result<Self, ValidationError>

Default implementation calls Self::validate.

The purpose of this validate function is to provide a place to run stricter validation, which is often wanted when creating things.